groq-chatbot/requirements.txt
streamlit
langchain
langchain-groq
python-dotenv
openai